requirements.txt
fitz==0.0.1.dev2
Flask==3.1.0
Pillow==11.0.0
